use clang-format style

+++ src/my/circle.cpp
@@ -10,6 +10,6 @@ MyCircle::MyCircle(float r, int n) {
 
   for (int i = 0; i < n; i++) {
     float angle = 2 * PI * i / n;
-    vertices[i].assign({ r *cos(angle), r * sin(angle) });
+    vertices[i].assign({r * cosf(angle), r * sinf(angle)});
   }
 }
